requirements.txt
asgiref==3.8.1
certifi==2024.6.2
charset-normalizer==3.3.2
dj-rest-auth==6.0.0
Django==4.2.6
django-rest-framework==0.1.0
djangorestframework==3.15.0
exceptiongroup==1.2.1
idna==3.7
iniconfig==2.0.0
packaging==24.1
pluggy==1.5.0
pytest==8.2.2
pytest-django==4.8.0
requests==2.32.3
sqlparse==0.5.0
tomli==2.0.1
typing_extensions==4.12.2
urllib3==1.26.6
